Settings: notification transparency (2/2)
this add back notification transparency...compared to 4.3
it controlls as well the notification icon backgound transparency
which felt in the past dissplaced. As well codecleanup of the old
implementation.

largeIcon support I did not add due that google does not support it
as well at current status...it is only a placeholder in
NotificationData